backend-api/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn==1.5.1
joblib
gunicorn
python-whois 
shap